Move `WindowEvent`, `PointerEventButton` and `Key` to the `platform` module
For `Key` this is a move, as it is a new type. For `WindowEvent` and `PointerEventButton` deprecated aliases
are provided.
