+++ tests/cases/elements/flickable_in_flickable.slint
@@ -32,7 +32,7 @@ TestCase := Window {
 /*
 
 ```rust
-use slint::{WindowEvent, PointerEventButton, LogicalPosition};
+use slint::{platform::WindowEvent, platform::PointerEventButton, LogicalPosition};
 let instance = TestCase::new();
 
 // First, try to scroll up
