+++ tests/cases/text/control_keys_input.slint
@@ -13,7 +13,7 @@ TestCase := Window {
 
 /*
 ```rust
-use slint::Key;
+use slint::platform::Key;
 
 let instance = TestCase::new();
 slint_testing::send_mouse_click(&instance, 5., 5.);
