+++ tests/cases/text/keyboard_modifiers.slint
@@ -30,7 +30,7 @@ TestCase := Window {
 
 /*
 ```rust
-use slint::Key;
+use slint::platform::Key;
 
 let instance = TestCase::new();
 slint_testing::send_mouse_click(&instance, 5., 5.);
